Jionghao temp4

+++ mysite/posts/views.py
@@ -26,14 +26,48 @@ from .serializers import PostSerializer, CommentSerializer, LikeSerializer
 from .authentication import UsernamePasswordAuthentication
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.decorators import authentication_classes
-from rest_framework.authentication import TokenAuthentication
+from rest_framework.authentication import TokenAuthentication, BaseAuthentication
 from django.http.response import JsonResponse
-from rest_framework.parsers import JSONParser 
+from rest_framework.parsers import JSONParser
 from rest_framework import status
 from rest_framework.authtoken.models import Token
 from rest_framework import exceptions
 import requests
 import re
+
+from rest_framework import authentication, permissions
+import base64
+
+
+class AccessPermission(permissions.BasePermission):
+    def has_permission(self, request, view):
+        auth_header = request.META.get('HTTP_AUTHORIZATION', '')
+        token_type, _, credentials = auth_header.partition(' ')
+
+        expected = base64.b64encode(b'socialdistribution_t05:c404t05').decode()
+        if token_type == 'Basic' and credentials == expected:
+            return True
+        elif token_type == 'Token':
+            return permissions.IsAuthenticated().has_permission(
+                request=request, view=view)
+        else:
+            return False
+
+
+class CustomAuthentication(authentication.BaseAuthentication):
+    def authenticate(self, request):
+        auth_header = request.META.get('HTTP_AUTHORIZATION', '')
+        token_type, _, credentials = auth_header.partition(' ')
+
+        expected = base64.b64encode(b'socialdistribution_t05:c404t05').decode()
+        if token_type == 'Basic' and credentials == expected:
+            return (True, None)
+
+        else:
+            return None
+
+    def authenticate_header(self, request):
+        return '{"username" : <username>, "password" : <password>}'
 
 
 # Create your views here.
@@ -118,22 +152,15 @@ def likePost(request, pk):
 
 
 @api_view(['GET'])
-@authentication_classes([TokenAuthentication])
-#@authentication_classes([UsernamePasswordAuthentication])
-@permission_classes([IsAuthenticated])
+@authentication_classes([CustomAuthentication])
+@permission_classes([AccessPermission])
 def request_post_list(request):
     posts = Post.objects.all()
     posts_serializer = PostSerializer(posts, many=True)
     return Response(posts_serializer.data)
 
 
-@api_view(['GET'])
-@authentication_classes([TokenAuthentication])
-@permission_classes([IsAuthenticated])
-def request_post(request, id):
-    post = Post.objects.get(id=id)
-    post_serializer = PostSerializer(post)
-    return Response(post_serializer.data)
+
 
 
 @api_view(['POST'])
@@ -156,11 +183,10 @@ def create_new_post(request):
                 {'message': 'The requested post does not exist'},
                 status=status.HTTP_404_NOT_FOUND)
 
-
-@api_view(['GET', 'POST', 'DELETE'])
-@authentication_classes([TokenAuthentication])
-@permission_classes([IsAuthenticated])
-def manage_user_post(request, username):
+@api_view(['GET'])
+@authentication_classes([CustomAuthentication])
+@permission_classes([AccessPermission])
+def get_posts_from_user(request, username):
     if request.method == 'GET':
         try:
             related_user = User.objects.get(username=username)
@@ -171,14 +197,23 @@ def manage_user_post(request, username):
             return JsonResponse(
                 {'message': 'The requested user does not exist'},
                 status=status.HTTP_404_NOT_FOUND)
+@api_view(['POST', 'DELETE'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
+def manage_user_post(request, username):
     if request.method == 'POST':
         try:
             related_user = User.objects.get(username=username)
 
-            user_token = Token.objects.get(user = related_user)
-            entered_token = re.findall('(?:Token\s)(\w*)', request.META['HTTP_AUTHORIZATION'])[0]
+            #This part of code will check if the token from request matches the token of author.
+            user_token = Token.objects.get(user=related_user)
+            entered_token = re.findall('(?:Token\s)(\w*)',
+                                       request.META['HTTP_AUTHORIZATION'])[0]
             if str(user_token) != entered_token:
-                return JsonResponse({'message': 'The token user does not match the post user'}, status = status.HTTP_403_FORBIDDEN)
+                return JsonResponse(
+                    {'message': 'The token user does not match the post user'},
+                    status=status.HTTP_403_FORBIDDEN)
+            #End of this part of code.
 
             data = JSONParser().parse(request)
             post_serializer = PostSerializer(data=data)
@@ -194,15 +229,20 @@ def manage_user_post(request, username):
                 status=status.HTTP_404_NOT_FOUND)
 
     if request.method == 'DELETE':
-        
+
         try:
             related_user = User.objects.get(username=username)
-            
-            user_token = Token.objects.get(user = related_user)
-            entered_token = re.findall('(?:Token\s)(\w*)', request.META['HTTP_AUTHORIZATION'])[0]
+
+            #This part of code will check if the token from request matches the token of author.
+            user_token = Token.objects.get(user=related_user)
+            entered_token = re.findall('(?:Token\s)(\w*)',
+                                       request.META['HTTP_AUTHORIZATION'])[0]
             if str(user_token) != entered_token:
-                return JsonResponse({'message': 'The token user does not match the post user'}, status = status.HTTP_403_FORBIDDEN)
-            
+                return JsonResponse(
+                    {'message': 'The token user does not match the post user'},
+                    status=status.HTTP_403_FORBIDDEN)
+            #End of this part of code.
+
             posts = Post.objects.filter(author=related_user)
             for post in posts:
                 post.delete()
@@ -213,8 +253,21 @@ def manage_user_post(request, username):
                 {'message': 'The requested user does not exist'},
                 status=status.HTTP_404_NOT_FOUND)
 
+@api_view(['GET'])
+@authentication_classes([CustomAuthentication])
+@permission_classes([AccessPermission])
+def request_post(request, id):
+    if request.method == 'GET':
+        try:
+            post = Post.objects.get(id=id)
+            post_serializer = PostSerializer(post)
+            return Response(post_serializer.data)
+        except Post.DoesNotExist:
+            return JsonResponse(
+                {'message': 'The requested post does not exist'},
+                status=status.HTTP_404_NOT_FOUND)
 
-@api_view(['GET', 'PUT', 'DELETE'])
+@api_view(['PUT', 'DELETE'])
 @authentication_classes([TokenAuthentication])
 @permission_classes([IsAuthenticated])
 def crud_post(request, id):
@@ -233,14 +286,16 @@ def crud_post(request, id):
     if request.method == 'PUT':
         try:
             related_post = Post.objects.get(id=id)
-            '''
-            related_user = related_post.author
-            print(related_user.username)
-            user_token = Token.objects.get(user = related_user)
-            entered_token = re.findall('(?:Token\s)(\w*)', request.META['HTTP_AUTHORIZATION'])[0]
+
+            #This part of code will check if the token from request matches the token of author.
+            user_token = Token.objects.get(user=related_post.author)
+            entered_token = re.findall('(?:Token\s)(\w*)',
+                                       request.META['HTTP_AUTHORIZATION'])[0]
             if str(user_token) != entered_token:
-                return JsonResponse({'message': 'The token user does not match the post user'}, status = status.HTTP_403_FORBIDDEN)
-            '''
+                return JsonResponse(
+                    {'message': 'The token user does not match the post user'},
+                    status=status.HTTP_403_FORBIDDEN)
+            #End of this part of code.
 
             data = JSONParser().parse(request)
             post_serializer = PostSerializer(related_post, data=data)
@@ -256,8 +311,19 @@ def crud_post(request, id):
 
     if request.method == 'DELETE':
         try:
-            post = Post.objects.get(id=id)
-            post.delete()
+            related_post = Post.objects.get(id=id)
+
+            #This part of code will check if the token from request matches the token of author.
+            user_token = Token.objects.get(user=related_post.author)
+            entered_token = re.findall('(?:Token\s)(\w*)',
+                                       request.META['HTTP_AUTHORIZATION'])[0]
+            if str(user_token) != entered_token:
+                return JsonResponse(
+                    {'message': 'The token user does not match the post user'},
+                    status=status.HTTP_403_FORBIDDEN)
+            #End of this part of code
+
+            related_post.delete()
             return JsonResponse({'message': 'Post was deleted'},
                                 status=status.HTTP_204_NO_CONTENT)
         except Post.DoesNotExist:
@@ -265,11 +331,10 @@ def crud_post(request, id):
                 {'message': 'The requested post does not exist'},
                 status=status.HTTP_404_NOT_FOUND)
 
-
-@api_view(['GET', 'POST', 'DELETE'])
-@authentication_classes([TokenAuthentication])
-@permission_classes([IsAuthenticated])
-def manage_post_comment(request, post_id):
+@api_view(['GET'])
+@authentication_classes([CustomAuthentication])
+@permission_classes([AccessPermission])
+def get_comments_from_post(request, post_id):
     if request.method == 'GET':
         try:
             related_post = Post.objects.get(id=post_id)
@@ -280,9 +345,26 @@ def manage_post_comment(request, post_id):
             return JsonResponse(
                 {'message': 'The requested post does not exist'},
                 status=status.HTTP_404_NOT_FOUND)
+
+@api_view(['POST', 'DELETE'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
+def manage_post_comment(request, post_id):
+ 
     if request.method == 'POST':
         try:
             related_post = Post.objects.get(id=post_id)
+
+            #This part of code will check if the token from request matches the token of author.
+            user_token = Token.objects.get(user=related_post.author)
+            entered_token = re.findall('(?:Token\s)(\w*)',
+                                       request.META['HTTP_AUTHORIZATION'])[0]
+            if str(user_token) != entered_token:
+                return JsonResponse(
+                    {'message': 'The token user does not match the post user'},
+                    status=status.HTTP_403_FORBIDDEN)
+            #End of this part of code.
+
             data = JSONParser().parse(request)
             comment_serializer = CommentSerializer(data=data)
             if comment_serializer.is_valid():
@@ -299,6 +381,17 @@ def manage_post_comment(request, post_id):
     if request.method == 'DELETE':
         try:
             related_post = Post.objects.get(id=post_id)
+
+            #This part of code will check if the token from request matches the token of author.
+            user_token = Token.objects.get(user=related_post.author)
+            entered_token = re.findall('(?:Token\s)(\w*)',
+                                       request.META['HTTP_AUTHORIZATION'])[0]
+            if str(user_token) != entered_token:
+                return JsonResponse(
+                    {'message': 'The token user does not match the post user'},
+                    status=status.HTTP_403_FORBIDDEN)
+            #End of this part of code.
+
             comments = Comment.objects.filter(post=related_post)
             for comment in comments:
                 comment.delete()
@@ -310,11 +403,10 @@ def manage_post_comment(request, post_id):
                 {'message': 'The requested post does not exist'},
                 status=status.HTTP_404_NOT_FOUND)
 
-
-@api_view(['GET', 'PUT', 'DELETE'])
-@authentication_classes([TokenAuthentication])
-@permission_classes([IsAuthenticated])
-def crud_comment(request, comment_id):
+@api_view(['GET'])
+@authentication_classes([CustomAuthentication])
+@permission_classes([AccessPermission])
+def request_comment(request, comment_id):
     if request.method == 'GET':
         try:
             related_comment = Comment.objects.get(id=comment_id)
@@ -324,6 +416,11 @@ def crud_comment(request, comment_id):
             return JsonResponse(
                 {'message': 'The requested comment does not exist'},
                 status=status.HTTP_404_NOT_FOUND)
+
+@api_view(['PUT', 'DELETE'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
+def crud_comment(request, comment_id):
     if request.method == 'PUT':
         try:
             related_comment = Comment.objects.get(id=comment_id)
@@ -348,21 +445,27 @@ def crud_comment(request, comment_id):
                 {'message': 'The requested comment does not exist'},
                 status=status.HTTP_404_NOT_FOUND)
 
-
-@api_view(['GET', 'POST', 'DELETE'])
-@authentication_classes([TokenAuthentication])
-@permission_classes([IsAuthenticated])
-def manage_post_like(request, post_id):
+@api_view(['GET'])
+@authentication_classes([CustomAuthentication])
+@permission_classes([AccessPermission])
+def get_likes_from_post(request, post_id):
     if request.method == 'GET':
         try:
             related_post = Post.objects.get(id=post_id)
             likes = Like.objects.filter(post=related_post)
             likes_serializer = LikeSerializer(likes, many=True)
+            print(likes_serializer.data)
             return JsonResponse(likes_serializer.data, safe=False)
         except Post.DoesNotExist:
             return JsonResponse(
                 {'message': 'The requested post does not exist'},
                 status=status.HTTP_404_NOT_FOUND)
+
+
+@api_view(['POST', 'DELETE'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
+def manage_post_like(request, post_id):
 
     if request.method == 'POST':
         try:
@@ -394,11 +497,10 @@ def manage_post_like(request, post_id):
                 {'message': 'The requested post does not exist'},
                 status=status.HTTP_404_NOT_FOUND)
 
-
-@api_view(['GET', 'PUT', 'DELETE'])
-@authentication_classes([TokenAuthentication])
-@permission_classes([IsAuthenticated])
-def crud_like(request, like_id):
+@api_view(['GET'])
+@authentication_classes([CustomAuthentication])
+@permission_classes([AccessPermission])
+def request_like(request, like_id):
     if request.method == 'GET':
         try:
             related_like = Like.objects.get(id=like_id)
@@ -408,6 +510,12 @@ def crud_like(request, like_id):
             return JsonResponse(
                 {'message': 'The requested like does not exist'},
                 status=status.HTTP_404_NOT_FOUND)
+
+@api_view(['PUT', 'DELETE'])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
+def crud_like(request, like_id):
+    
     if request.method == 'PUT':
         try:
             related_like = Like.objects.get(id=like_id)
@@ -520,29 +628,21 @@ class SharedPostView(View):
             shared_user=current_user,
             contentType=post_object.contentType).save()
 
+
 def send_token(request, username, password):
-    
-    
+
     data = {'username': username, 'password': password}
-    
-    response = requests.post('http://127.0.0.1:8000/api-token-auth/', json = data)
+
+    response = requests.post('https://cmput404-socialdist-project.herokuapp.com/api-token-auth/', json=data)
     dict_data = ast.literal_eval(response.text)
     print(ast.literal_eval(response.text))
     return JsonResponse(dict_data, safe=False)
     
-    '''
-    user = User.objects.get(username=username)
-    token = Token.objects.get(user=user)
-    user_password = user.password
-    if user_password == password:
-        #return Response(token.key, safe=False)
-    
-    
-        print(token)
-    return {'message': 'The user or password does not exist'}
-    '''
+
+
 #curl -X POST -d "username=1&password=12345" http://127.0.0.1:8000/api-token-auth/
 #curl -X POST -d '{"title":"This is a new post","text":"a new post is here","image":null,"pub_date":"2021-11-09T21:51:55.850726Z","author":2,"shared_user":null,"shared_on":null,"privacy":0,"contentType":"text/plain"}' http://127.0.0.1:8000/post/create_new_post -H 'Authorization: Token 8a91340fa2849cdc7e0e7aa07f4b2c0e91f09a3a'
 #curl -X GET http://127.0.0.1:8000/post/manage_user_post/1 -H 'Authorization: Token 8a91340fa2849cdc7e0e7aa07f4b2c0e91f09a3a'
 #curl -X POST -d '{"title":"This is a new post","text":"a new post is here","image":null,"pub_date":"2021-11-09T21:51:55.850726Z","author":2,"shared_user":null,"shared_on":null,"privacy":0,"contentType":"text/plain"}' http://127.0.0.1:8000/post/manage_user_post/9 -H 'Authorization: Token 8a91340fa2849cdc7e0e7aa07f4b2c0e91f09a3a'
 #curl -X PUT -d '{"title":"This is a new post","text":"a new post is here","image":null,"pub_date":"2021-11-09T21:51:55.850726Z","author":2,"shared_user":null,"shared_on":null,"privacy":0,"contentType":"text/plain"}' http://127.0.0.1:8000/post/crud_post/42 -H 'Authorization: Token 8a91340fa2849cdc7e0e7aa07f4b2c0e91f09a3a'
+#curl -X DELETE http://127.0.0.1:8000/post/crud_post/33 -H 'Authorization: Token 8a91340fa2849cdc7e0e7aa07f4b2c0e91f09a3a'
